# src/tools/platform-api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exalto Seal API",
    "description": "Account, Device, hosted Trace, public Trace, Registry, verification, admission, usage, credits, and billing API for Exalto Seal. This contract is separate from the loopback local administration API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.exalto.ai",
      "description": "Exalto Seal"
    }
  ],
  "paths": {
    "/api/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get the current Exalto account",
        "operationId": "account",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete the current Exalto account and hosted data",
        "operationId": "delete_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account deleted",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
    "/api/api-keys": {
      "get": {
        "tags": [
          "api-keys"
        ],
        "summary": "List account API keys",
        "operationId": "list_api_keys",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; defaults to 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "Create an account API key",
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/api/auth/providers": {
      "get": {
        "tags": [
          "browser-auth"
        ],
        "summary": "List configured browser sign-in providers",
        "operationId": "auth_providers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthProvidersResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          }
        }
      }
    },
...
    "/api/billing/purchases": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List recent hosted-credit purchases",
        "operationId": "list_purchases",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPurchasesResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/api/credit-offers": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "List promotional credit offers eligible for the signed-in account",
        "operationId": "eligible_credit_offers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditOffersResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/api/credits/history": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "List the signed-in account's credit activity",
        "operationId": "credit_history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; defaults to 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CreditHistoryEntry"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/api/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "List the account's Connected devices",
        "operationId": "list_devices",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; defaults to 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ConnectedDevice"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/api/public/traces": {
      "get": {
        "tags": [
          "public-traces"
        ],
        "summary": "List Listed verified-session traces",
        "operationId": "list_public_traces",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; defaults to 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shared_after",
            "in": "query",
            "description": "Include traces shared at or after this Unix timestamp",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PublicTraceSummary"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/traces/{trace_id}/package.llmtrace": {
      "get": {
        "tags": [
          "public-traces"
        ],
        "summary": "Download the exact verified portable proof package",
        "operationId": "public_trace_package",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/vnd.exalto.notary.trace-package+zip": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/traces/{trace_id}/trace.otlp.json": {
      "get": {
        "tags": [
          "public-traces"
        ],
        "summary": "Download the verified canonical OpenTelemetry trace",
        "operationId": "public_trace_otlp",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/registry": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get the versioned Registry of Official Notaries",
        "operationId": "get_registry",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryResponse"
                }
              }
            }
          },
          "403": {
            "description": "Browser origin denied (browser_origin_denied)"
          }
        }
      }
    },
...
    "/api/usage": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get capture, notarization, and hosted Trace usage",
        "operationId": "usage",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          },
          {
            "bearerAuth": [
              "account:read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "type": "object",
        "required": [
          "account",
          "billing",
          "links"
        ],
        "properties": {
          "account": {
            "$ref": "#/components/schemas/PublicUser"
          },
          "billing": {
            "$ref": "#/components/schemas/AccountBillingResponse"
          },
          "links": {
            "$ref": "#/components/schemas/AccountLinks"
          }
        }
      },
...
      "AuthProvidersResponse": {
        "type": "object",
        "required": [
          "github",
          "google"
        ],
        "properties": {
          "github": {
            "type": "boolean"
          },
          "google": {
            "type": "boolean"
          }
        }
      },
...
      "BillingPurchasesResponse": {
        "type": "object",
        "required": [
          "purchases"
        ],
        "properties": {
          "purchases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingPurchase"
            }
          }
        }
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": [
          "name",
          "scopes"
        ],
        "properties": {
          "expires_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CreateApiKeyResponse": {
        "type": "object",
        "required": [
          "api_key",
          "secret"
        ],
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/ApiKeyResponse"
          },
          "secret": {
            "type": "string"
          }
        }
      },
...
      "CreditOffersResponse": {
        "type": "object",
        "required": [
          "offers"
        ],
        "properties": {
          "offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditOffer"
            }
          }
        }
      },
...
      "DeleteAccountRequest": {
        "type": "object",
        "required": [
          "confirmation"
        ],
        "properties": {
          "confirmation": {
            "type": "string",
            "description": "Must be exactly `DELETE` after the user confirms local traces remain local."
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Page_ApiKeyResponse": {
        "type": "object",
        "description": "Uniform response shape for every paginated list route.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "prefix",
                "name",
                "scopes",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "integer",
                  "format": "int64"
                },
                "expires_at": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "id": {
                  "type": "string"
                },
                "last_used_at": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "name": {
                  "type": "string"
                },
                "prefix": {
                  "type": "string"
                },
                "revoked_at": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "scopes": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiScope"
                  }
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cursor for the next page, or `null` when this page exhausts the query."
          }
        }
      },
...
      "RegistryResponse": {
        "type": "object",
        "required": [
          "format",
          "generation",
          "active_key_id",
          "notaries"
        ],
        "properties": {
          "active_key_id": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "generation": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "notaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistryRecordResponse"
            }
          }
        },
        "additionalProperties": false
      },
...
      "UsageResponse": {
        "type": "object",
        "required": [
          "credits",
          "operations",
          "hosted_traces"
        ],
        "properties": {
          "credits": {
            "$ref": "#/components/schemas/CreditSummary"
          },
          "hosted_traces": {
            "$ref": "#/components/schemas/HostedTraceUsage"
          },
          "operations": {
            "$ref": "#/components/schemas/NotaryStats"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Short-lived Device access token or stable scoped API key"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "notary_session",
        "description": "HttpOnly hosted browser session cookie"
      },
      "pollSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Notary-Poll-Secret",
        "description": "One-time secret returned when Device authorization starts"
      },
      "serviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Dedicated notary-to-platform service credential"
      }
    }
  },
...
}